@@ -13,32 +13,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/cli": "1.2.0",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/language-service": "^4.0.0",
-    "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
-    "body-parser": "^1.17.2",
-    "commander": "^2.11.0",
-    "compression": "^1.7.0",
-    "core-js": "^2.4.1",
-    "express": "^4.15.3",
+    "@angular/animations": "^5.2.10",
+    "@angular/cdk": "^5.2.5",
+    "@angular/cli": "1.7.4",
+    "@angular/common": "^5.2.10",
+    "@angular/compiler": "^5.2.10",
+    "@angular/compiler-cli": "^5.2.10",
+    "@angular/core": "^5.2.10",
+    "@angular/forms": "^5.2.10",
+    "@angular/http": "^5.2.10",
+    "@angular/language-service": "^5.2.10",
+    "@angular/material": "^5.2.5",
+    "@angular/material-moment-adapter": "^5.2.5",
+    "@angular/platform-browser": "^5.2.10",
+    "@angular/platform-browser-dynamic": "^5.2.10",
+    "@angular/router": "^5.2.10",
+    "body-parser": "^1.18.2",
+    "commander": "^2.15.1",
+    "compression": "^1.7.2",
+    "core-js": "^2.5.5",
+    "express": "^4.16.3",
     "express-history-api-fallback": "^2.2.1",
     "google-spreadsheet": "^2.0.4",
-    "moment": "^2.18.1",
-    "moment-timezone": "^0.5.13",
-    "rxjs": "^5.1.0",
-    "typescript": "^2.4.1",
-    "webpack-dev-middleware": "^1.11.0",
-    "zone.js": "^0.8.4"
+    "moment": "^2.22.1",
+    "moment-timezone": "^0.5.16",
+    "rxjs": "^5.0.0",
+    "typescript": ">=2.4.2 <2.7.0",
+    "webpack-dev-middleware": "^3.1.3",
+    "zone.js": "^0.8.26"
   }
 }
